requirements.txt
certifi==2019.11.28
chardet==3.0.4
Django==2.2.7
djangorestframework==3.10.3
idna==2.8
psycopg2-binary==2.8.4
Pygments==2.5.2
pytz==2019.3
requests==2.22.0
sqlparse==0.3.0
urllib3==1.25.7
